+++ backend/src/main/java/com/coffee/backend/dto/response/ProductoListadoResponseDTO.java
@@ -1,8 +1,10 @@
 package com.coffee.backend.dto.response;
 
+import java.math.BigDecimal;
+
 public record ProductoListadoResponseDTO(Long productoId,
                                          String nombreProducto,
                                          String categoria,
-                                         Double precio,
+                                         BigDecimal precio,
                                          Boolean disponible) {
 }
