+++ backend/src/main/java/com/coffee/backend/dto/response/ProductosDisponiblesResponseDTO.java
@@ -1,9 +1,11 @@
 package com.coffee.backend.dto.response;
 
+import java.math.BigDecimal;
+
 public record ProductosDisponiblesResponseDTO(
       Long productoId,
       String nombreProducto,
-      Double precio,
+      BigDecimal precio,
       String categoria,
       Boolean disponible
 ) {
